Remove targetSDK
Remove settings target sdk from the manifest as it should target the
system it is compiled with similar to phone settings. Android 35 expects
apps to be edge-to-edge which creates some issues with keyboard
resizing, so some activities needed to be updated so their content
doesn't get hidden by the keyboard.

Bug: 360700047
Test: Manual and atest android.platform.tests.SecuritySettingTest
Flag: EXEMPT manifest change
Change-Id: I1d13f14ebf528f2cbe71679a6f54f029ef48a08f